refactor: external the logical in SortableManager for lifecycle management of sortable instances
